// src/lib/types.ts
export type AnswerSelectionType = 'single' | 'multiple';
// ...
export interface ButtonState {
  [key: number]: {
    disabled?: boolean;
    className?: string;
    selected?: boolean;
  };
}

export interface CheckAnswerParams {
  userInput: any[];
  userAttempt: number;
  currentQuestionIndex: number;
  continueTillCorrect?: boolean;
  showNextQuestionButton: boolean;
  incorrect: number[];
  correct: number[];
  setButtons: React.Dispatch<React.SetStateAction<ButtonState>>;
  setIsCorrect: React.Dispatch<React.SetStateAction<boolean>>;
  setIncorrectAnswer: React.Dispatch<React.SetStateAction<boolean>>;
  setCorrect: React.Dispatch<React.SetStateAction<number[]>>;
  setIncorrect: React.Dispatch<React.SetStateAction<number[]>>;
  setShowNextQuestionButton: React.Dispatch<React.SetStateAction<boolean>>;
  setUserInput: React.Dispatch<React.SetStateAction<any[]>>;
  setUserAttempt: React.Dispatch<React.SetStateAction<number>>;
}

export interface SelectAnswerParams {
  userInput: any[];
  currentQuestionIndex: number;
  setButtons: React.Dispatch<React.SetStateAction<ButtonState>>;
  setShowNextQuestionButton: React.Dispatch<React.SetStateAction<boolean>>;
  incorrect: number[];
  correct: number[];
  setCorrect: React.Dispatch<React.SetStateAction<number[]>>;
  setIncorrect: React.Dispatch<React.SetStateAction<number[]>>;
  setUserInput: React.Dispatch<React.SetStateAction<any[]>>;
}